pyproject.toml
[tool.poetry]
name = "blog-will"
version = "0.1.0"
description = ""
authors = ["Ulrich Wilfried Motchoffo Kom <motchoffow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
httpie = "^2.3.0"
djangorestframework-simplejwt = "^4.6.0"
Django = "^3.1.4"

[tool.poetry.dev-dependencies]
Django = "^3.1.4"
djangorestframework = "^3.12.2"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
